requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
